Encode/decode MtuProbingCompleteSupport::Enabled
Previously the implementation of MtuProbingCompleteSupport encoded it to
the wire if it was disabled, which isn't the right behavior. This is
probably a copy/paste mistake from the MigrationSupport parameter which
on the wire is `disable_active_migration` and so needs to get encoded
when disabled. MtuProbingCompleteSupport is encoded when enabled
instead.